fix syntax

+++ storage/src/stream.cc
@@ -757,7 +757,7 @@ unsigned int stream::_find_metric_id(
       _update_metrics.bind_value(":warn", check_double(warn));
       _update_metrics.bind_value(":warn_low", check_double(warn_low));
       _update_metrics.bind_value(":warn_threshold_mode", warn_mode);
-      _update_metrics.bind_value(":crit", check_double(crit.toFloat()));
+      _update_metrics.bind_value(":crit", check_double(crit));
       _update_metrics.bind_value(":crit_low", check_double(crit_low));
       _update_metrics.bind_value(":crit_threshold_mode", crit_mode);
       _update_metrics.bind_value(":min", check_double(min));
